# apps/luxledger-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LuxLedger API
  version: 0.2.0
  description: Reference REST API built on top of @lux/ledger.
servers:
  - url: http://localhost:3000
security:
  - BearerAuth: []
paths:
...
  /v1/transactions:
    post:
      summary: Create transaction (idempotent by tenant+reference)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTransactionResponse'
        '200':
          description: Idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    get:
      summary: List transactions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/LedgerIdQuery'
      responses:
        '200':
          description: Transaction page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsPage'
              examples:
                default:
                  value:
                    data:
                      - id: 00000000-0000-4000-8000-000000000201
                        tenant_id: 11111111-1111-4111-8111-111111111111
                        ledger_id: 00000000-0000-4000-8000-000000000001
                        reference: tx-ref-1
                        currency: USD
                        description: Payment settlement
                        related_transaction_id: null
                        relation_type: null
                        effective_at: '2026-01-01T00:01:00.000Z'
                        created_at: '2026-01-01T00:01:00.000Z'
                    next_cursor: eyJjcmVhdGVkX2F0IjoiMjAyNi0wMS0wMVQwMDowMTowMC4wMDBaIiwiaWQiOiIwMDAwMDAwMC0wMDAwLTQwMDAtODAwMC0wMDAwMDAwMDAyMDEifQ
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /v1/transactions/{id}:
    get:
      summary: Get transaction by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
              examples:
                found:
                  value:
                    id: 00000000-0000-4000-8000-000000000201
                    tenant_id: 11111111-1111-4111-8111-111111111111
                    ledger_id: 00000000-0000-4000-8000-000000000001
                    reference: tx-ref-1
                    currency: USD
                    description: Payment settlement
                    related_transaction_id: null
                    relation_type: null
                    created_at: '2026-01-01T00:01:00.000Z'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /v1/transactions/{id}/reverse:
    post:
      summary: Reverse transaction (idempotent by tenant+reference)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReverseTransactionRequest'
            examples:
              default:
                value:
                  reference: tx-ref-1-reversal
                  description: Reverse invalid posting
      responses:
        '201':
          description: Reversal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTransactionResponse'
        '200':
          description: Idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTransactionResponse'
  /v1/transactions/{id}/correct:
    post:
      summary: Correct transaction by creating reversal + replacement (atomic, idempotent)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectTransactionRequest'
            examples:
              default:
                value:
                  reversal_reference: tx-ref-1-reversal
                  corrected_reference: tx-ref-1-corrected
                  description: Correct account allocation
                  entries:
                    - account_id: 00000000-0000-4000-8000-000000000101
                      direction: DEBIT
                      amount_minor: '100'
                      currency: USD
                    - account_id: 00000000-0000-4000-8000-000000000102
                      direction: CREDIT
                      amount_minor: '100'
                      currency: USD
      responses:
        '201':
          description: Correction flow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectTransactionResponse'
        '200':
          description: Idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectTransactionResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      required: false
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
      required: false
    LedgerIdQuery:
      name: ledger_id
      in: query
      schema:
        type: string
        format: uuid
      required: false
...
  responses:
    BadRequestError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    TransactionEntryRequest:
      type: object
      additionalProperties: false
      required: [account_id, direction, amount_minor, currency]
      properties:
        account_id:
          type: string
          format: uuid
        direction:
          type: string
          enum: [DEBIT, CREDIT]
        amount_minor:
          type: string
          pattern: '^[1-9][0-9]*$'
        currency:
          type: string
          pattern: '^(?=.*\S).+$'
    CreateTransactionRequest:
      type: object
      additionalProperties: false
      required: [ledger_id, reference, currency, entries]
      properties:
        ledger_id:
          type: string
          format: uuid
        reference:
          type: string
          pattern: '^(?=.*\S).+$'
        currency:
          type: string
          pattern: '^(?=.*\S).+$'
        description:
          type: string
          pattern: '^(?=.*\S).+$'
          description: Optional transaction description. When omitted it is stored as null.
        effective_at:
          type: string
          format: date-time
          description: Optional accounting effective timestamp for backdated postings.
        entries:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/TransactionEntryRequest'
...
    CreateTransactionResponse:
      type: object
      additionalProperties: false
      required: [transaction_id, created]
      properties:
        transaction_id:
          type: string
          format: uuid
        created:
          type: boolean
...
    CorrectTransactionResponse:
      type: object
      additionalProperties: false
      required: [reversal_transaction_id, corrected_transaction_id, created]
      properties:
        reversal_transaction_id:
          type: string
          format: uuid
        corrected_transaction_id:
          type: string
          format: uuid
        created:
          type: boolean
    ReverseTransactionRequest:
      type: object
      additionalProperties: false
      required: [reference]
      properties:
        reference:
          type: string
          pattern: '^(?=.*\S).+$'
        description:
          type: string
          pattern: '^(?=.*\S).+$'
    CorrectTransactionRequest:
      type: object
      additionalProperties: false
      required: [reversal_reference, corrected_reference, entries]
      properties:
        reversal_reference:
          type: string
          pattern: '^(?=.*\S).+$'
        corrected_reference:
          type: string
          pattern: '^(?=.*\S).+$'
        description:
          type: string
          pattern: '^(?=.*\S).+$'
        entries:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/TransactionEntryRequest'
...
    TransactionsPage:
      type: object
      additionalProperties: false
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        next_cursor:
          type: string
          nullable: true
    Transaction:
      type: object
      additionalProperties: false
      required: [id, tenant_id, ledger_id, reference, currency, description, related_transaction_id, relation_type, effective_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        ledger_id:
          type: string
          format: uuid
        reference:
          type: string
        currency:
          type: string
        description:
          type: string
          nullable: true
        related_transaction_id:
          type: string
          format: uuid
          nullable: true
        relation_type:
          type: string
          enum: [REVERSAL, CORRECTION]
          nullable: true
        effective_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      example:
        id: 00000000-0000-4000-8000-000000000201
        tenant_id: 11111111-1111-4111-8111-111111111111
        ledger_id: 00000000-0000-4000-8000-000000000001
        reference: tx-ref-1
        currency: USD
        description: Payment settlement
        related_transaction_id: null
        relation_type: null
        effective_at: '2026-01-01T00:01:00.000Z'
        created_at: '2026-01-01T00:01:00.000Z'
...